// src/front/src/api/identity/models/appUserListFilter.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Identity API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { Department } from "./department";

export interface AppUserListFilter {
	pageNumber?: number;
	pageSize?: number;
	readonly skip?: number;
	/** @nullable */
	search?: string | null;
	/** @nullable */
	isActive?: boolean | null;
	department?: Department;
	/** @nullable */
	structureId?: number | null;
	/** @nullable */
	teamId?: number | null;
}